backend/requirements.txt
# requirements.txt - Main project dependencies
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
Pillow==10.1.0
drf-spectacular==0.26.4  # OpenAPI 3 schema generation

# Production-specific dependencies
gunicorn==21.2.0
whitenoise==6.6.0

# requirements-dev.txt - Development-specific dependencies
-r requirements.txt
flake8==6.1.0
black==23.10.1
isort==5.12.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
model-bakery==1.17.0

# Optional but recommended security and performance tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3